finish

+++ admin/views/software.php
@@ -1,11 +1,14 @@
 <!--Show On Index-->
 <?php
 $path_basemodel                 = "models/BaseModel.php";
-$path_software                  = "../../assets/software/";
-$path_manual                    = "../../assets/manual/";
+$path_software                  = "../assets/software/";
+$path_manual                    = "../assets/manual/";
+$path_manualupload              = "views/softwaremanual.php";
 $path_sofwaremodel_upload       = "models/softwaremodel/softwareupload.php";
 $path_sofwaremodel_delete       = "models/softwaremodel/softwaredelete.php";
-$path_sofwaremodel_download     = "models/softwaremodel/softwaredownload.php";
+$path_sofwaremodel_manual       = "models/softwaremodel/manualupload.php";
+
+
 
 ?>
 
@@ -23,13 +26,6 @@ $path_sofwaremodel_download     = "models/softwaremodel/softwaredownload.php";
                 <div class="col-sm-8" align="left">
                     ชื่อไฟล์
                     <input class="form-control" type="text" name="software_name" id="software_name" value="" required>
-                </div>
-            </div>
-            <div class="form-group">
-                <div class="col-sm-2" align="right"></div>
-                <div class="col-sm-8" align="left">
-                    รายละเอียดไฟล์
-                    <textarea class="form-control" name="software_desription_en" id="software_desription_en"></textarea>
                 </div>
             </div>
             <div class="form-group">
@@ -74,7 +70,7 @@ $path_sofwaremodel_download     = "models/softwaremodel/softwaredownload.php";
                 <div class="col-sm-2" align="right"></div>
                 <div class="col-sm-7" align="left">
                     ไฟล์ประกอบ
-                    <input type="file" name="fileupload" id="fileupload">
+                    <input type="file" name="fileupload" id="fileupload" require>
                 </div>
             </div>
             <div class="form-group">
@@ -88,16 +84,16 @@ $path_sofwaremodel_download     = "models/softwaremodel/softwaredownload.php";
 </div>
 
 <!--Show Detail Software-->
-<div class="col-xs-12 col-sm-12 col-md-12">
+<div class="col-xs-12 col-sm-12 col-md-12" id="manual">
     <div class="thumbnail" style=" width:100%; height:150px; overflow: auto;">
         <form>
-            <table border="0" cellpadding="3" cellspacing="1" width="100%">
+            <table cellpadding="3" cellspacing="1" width="100%">
                 <tr>
                     <td align="center"><strong>File Name</strong></td>
-                    <td align="center"><strong>Description</strong></td>
+                    <td align="center"><strong>File Manual</strong></td>
                     <td align="center"><strong>Type</strong></td>
                     <td align="center"><strong>Category</strong></td>
-                    <td align="center" colspan="2"><strong>Option</strong></td>
+                    <td align="center" colspan="3"><strong>Option</strong></td>
                 </tr>
                 <?php
                 include $path_basemodel;
@@ -112,15 +108,31 @@ $path_sofwaremodel_download     = "models/softwaremodel/softwaredownload.php";
                 ?>
                     <tr>
                         <td align="center"><?php echo $row["software_name"]; ?></td>
-                        <td align="center"><?php echo $row["software_description_en"]; ?></td>
+                        <td align="center"><?php
+                                            if ($row["software_manual"] != "") {
+                                                echo "มีไฟล์";
+                                            } else { ?>
+                                <a href="<?php echo $path_manualupload; ?>?software_id=<?php echo $row["software_id"]; ?>" onclick="return confirm('ต้องการเพิ่มข้อมูลหรือไม่')">upload</a>
+
+                            <?php
+                                            }
+                            ?></td>
                         <td align="center"><?php echo $row["software_type_name"]; ?></td>
                         <td align="center"><?php echo $row["product_type_name"]; ?></td>
                         <td align="center">
-                            <a href="<?php echo $path_software.$row["software_file"]; ?>" download><i class="fas fa-download"></i></a>
+                            <a href="<?php echo $path_software . $row["software_file"]; ?>" download><i class="fas fa-download"></i></a>
                         </td>
                         <td align="center">
-                            <a href="<?php echo $path_sofwaremodel_delete; ?>?delete_id=<?php echo $row["software_id"]; ?>&delete_file=<?php echo $row["software_file"]; ?>" 
-                            onclick="return confirm('ต้องการลบข้อมูลหรือไม่')"><i class="fa fa-trash" aria-hidden="true"></i></a>
+                            <?php if ($row["software_manual"] != "") { ?>
+                                <a href="<?php echo $path_manual . $row["software_manual"]; ?>" download><i class="fas fa-file-download"></i></a>
+                            <?php
+                            } else { ?>
+                                <i class="fas fa-ban"></i>
+                            <?php
+                            } ?>
+                        </td>
+                        <td align="center">
+                            <a href="<?php echo $path_sofwaremodel_delete; ?>?delete_id=<?php echo $row["software_id"]; ?>&delete_file=<?php echo $row["software_file"]; ?>" onclick="return confirm('ต้องการลบข้อมูลหรือไม่')"><i class="fa fa-trash" aria-hidden="true"></i></a>
                         </td>
                     </tr>
                 <?php
